+++ src/attribute/attributes.rs
@@ -157,8 +157,8 @@ impl Attributes {
     /// [`Node::set_attribute()`]: type.Node.html#method.set_attribute
     pub fn insert(&mut self, attr: Attribute) {
         if cfg!(debug_assertions) {
-            if attr.is_link() || attr.is_func_link() {
-                panic!("attribute with Link/FuncLink value must be set only via Node::set_attribute");
+            if attr.is_link() || attr.is_func_link() || attr.is_paint() {
+                panic!("attribute with Link/FuncLink/Paint value must be set only via Node::set_attribute");
             }
         }
 
@@ -213,8 +213,8 @@ impl Attributes {
             let name = AttributeQNameRef::from(name);
             let attr = self.0.iter().find(|x| x.name.as_ref() == name);
             if let Some(attr) = attr {
-                if attr.is_link() || attr.is_func_link() {
-                    panic!("attribute with Link/FuncLink value must be removed \
+                if attr.is_link() || attr.is_func_link() || attr.is_paint() {
+                    panic!("attribute with Link/FuncLink/Paint value must be removed \
                             only via Node::remove_attribute");
                 }
             }
@@ -288,8 +288,8 @@ impl Attributes {
         if cfg!(debug_assertions) {
             for attr in &self.0 {
                 if !f(attr) {
-                    if attr.is_link() || attr.is_func_link() {
-                        panic!("attribute with Link/FuncLink value must be removed \
+                    if attr.is_link() || attr.is_func_link() || attr.is_paint() {
+                        panic!("attribute with Link/FuncLink/Paint value must be removed \
                                 only via Node::remove_attribute");
                     }
                 }